# fe/vdbas_exp_fe/apiContract/API-contract.yaml
openapi: 3.0.3
info:
  title: EXP CAPEX Dossier API
  description: |
    API quản lý **Hồ sơ Chi đầu tư (CAPEX)** – Phân hệ Chi (EXP) trong hệ thống VDBAS.

    ## Luồng trạng thái hồ sơ
    ```
    Start → DRAFT → SAVED → SUBMITTED → APPROVED (Checker) → APPROVED (Approver) → COMPLETED
                                ↘ REJECTED (Checker reject / Approver reject)
               ↘ CANCELLED (Maker xoá)
    ```

    ## Phân quyền
    - **Maker**: Tạo mới, Sửa, Xoá, Lưu nháp, Gửi kiểm soát (trong phạm vi TREASURY_CODE)
    - **Checker**: Kiểm soát (Approve/Reject) hồ sơ ở trạng thái SUBMITTED
    - **Approver**: Phê duyệt (Approve/Reject) hồ sơ ở trạng thái APPROVED (đã kiểm soát)
    - **Viewer**: Chỉ xem

    ## Lưu ý kỹ thuật
    - Mọi ID kiểu UUID (chuỗi 36 ký tự, dạng `xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx`)
    - Ngày dùng `yyyy-MM-dd`; DateTime dùng ISO 8601 `yyyy-MM-ddTHH:mm:ss+07:00`
    - Số tiền VND kiểu `integer` (đơn vị VND)
    - Optimistic lock qua trường `version` (gửi kèm khi Update)
    - Phòng double-submit: header `X-Idempotency-Key` (UUID do client sinh)
    - Phân quyền theo `TREASURY_CODE` lấy từ JWT claim
  version: "0.2.0"
  contact:
    name: VDBAS Development Team
servers:
  - url: /api/v1
    description: API version 1
security:
  - bearerAuth: []
...
paths:
...
  /lov/projects:
    get:
      tags: [LOV]
      operationId: lovProjects
      summary: LOV.01 – Tra cứu Dự án/Công trình
      description: |
        Hỗ trợ F4 Lookup tra cứu dự án. Lọc theo mã, tên, loại, ban QLDA.
        Chỉ trả về dự án STATUS=1 (đang hiệu lực).
      parameters:
        - name: projectCode
          in: query
          schema: { type: string }
        - name: projectName
          in: query
          schema: { type: string }
        - name: projectType
          in: query
          schema: { type: string, enum: [MILITARY, CITIZEN] }
        - name: projectManagementCode
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Danh sách dự án
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectLovResponse'
  /lov/projects/{projectCode}/specific:
    get:
      tags: [LOV]
      operationId: lovProjectSpecific
      summary: LOV – Dự án đặc thù theo dự án cha (Military only)
      parameters:
        - name: projectCode
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Danh sách dự án đặc thù
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSpecificLovResponse'
  /lov/treasuries:
    get:
      tags: [LOV]
      operationId: lovTreasuries
      summary: LOV.04 – Danh mục Kho bạc
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Danh sách kho bạc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreasuryLovResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT chứa claims userId, role, treasuryCode
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalRecords:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
...
    ProjectLovItem:
      type: object
      properties:
        projectCode:
          type: string
          example: "7122155"
        projectName:
          type: string
          example: "Dự án nâng cấp bệnh viện Bạch Mai"
        projectType:
          type: string
          enum: [MILITARY, CITIZEN]
        projectManagementCode:
          type: string
          example: "3029123"
        projectManagementName:
          type: string
          example: "BQLDA bệnh viện Bạch Mai"
        investorCode:
          type: string
        investorName:
          type: string
        hasSpecific:
          type: boolean
          description: "true nếu loại MILITARY và có dự án đặc thù"
    ProjectLovResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ProjectLovItem'
            pagination:
              $ref: '#/components/schemas/Pagination'
    ProjectSpecificLovItem:
      type: object
      properties:
        projectSpecificCode:
          type: string
          example: "001200037"
        projectSpecificName:
          type: string
          example: "Dự án TM02"
        projectCode:
          type: string
    ProjectSpecificLovResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSpecificLovItem'
    TreasuryLovResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            properties:
              treasuryCode: { type: string, example: "0012" }
              treasuryName: { type: string, example: "Kho bạc nhà nước khu vực I - PGD số 12" }
...